python-requirements.txt
#openshift==0.12.1
#kubernetes==12.0.0
#ansible==5.7.1
#pywinrm==0.4.2
#jmespath==0.10.0
#jsondiff==1.3.1
#docker-compose==1.29.2
#cryptography==36.0.2 # https://github.com/paramiko/paramiko/issues/2038
#hcloud==1.16.0
#netaddr==0.8.0
jsonpatch # 1.32
PyYAML #5.4.1
kubernetes # 24.2.0
cffi #1.15.1
openshift # 0.13.1
typer
rich
shellingham
loguru
ansible
pywinrm
jmespath
jsondiff
docker-compose
cryptography
hcloud
netaddr
boto3
#azure-cli
python-benedict